requirements.txt
asttokens==3.0.1
boto3==1.42.23
botocore==1.42.23
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
dill==0.4.0
drake==1.48.0
executing==2.2.1
Farama-Notifications==0.0.4
fonttools==4.61.1
grpcio==1.76.0
gymnasium==1.2.3
h5py==3.15.1
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
ipdb==0.13.13
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.0.1
kiwisolver==1.4.9
matplotlib==3.10.8
matplotlib-inline==0.2.1
Mosek==11.0.24
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydot==4.0.1
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
s3transfer==0.16.0
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.2.14
wheel==0.45.1
